[스프링 - 협업 미션] 손너잘(손민성) 미션 제출합니다. (#38)
* style: sql 테이블명 소문자로 수정

* feat: mysql 의존 추가

* feat: mysql 드라이버 설정 추가

* feat: mysql username, password 수정

* feat: 첫 실행시에만 더미 데이터 생성하도록 구현

* fix: application-local.properties에 잘못된 account name 제거

* feat: production의 데이터베이스 ip 수정

* fix: production의 데이터베이스 private ip로 수정

* fix: production의 데이터베이스 private ip로 수정

* feat: datasource schema 추가

* feat: 사용되지 않는 application.yml 제거

* feat: datasource initialization mode 활성화

* refactor: 패키지 구조 변경

* feat: 구간 정보 조회 기능 구현

협업을 위한 요구사항

* test: 지하철 노선 내의 구간 조회 TC 작성

* refactor: 메소드 분리 및 불변객체화

* feat: 역 수정 기능 구현

협업을 위한 요구사항

* test: station update 기능 테스트 추

* feat: 구간 수정 기능 구현

* test: 구간 수정 기능 TC 작성

* fix: 구간 수정시 거리가 변경되지 않는 버그 해결

* feat: 지하철 역 목록 조회 구현

* test: 역 목록 조회 테스트

* feat: 전체보기기능 구현

* fix: 전체보기기능 요구사항과 다른 response 수정

* test: 전체 노선도 보기 TC 작성

* fix: StationOfMapResponse에 getDistance 추가

* test: 요구사항 변경으로 인한 TC 수정

* fix: 테이블명이 매칭이되지 않는 문제 해결

* refactor: 전체적인 컨벤션 및 코드 정리

* refactor: dto 필드명 수정

* refactor: 트랜잭션 추가

* refactor: 예외처리 로직 변경

커스텀 예외 추가
advice 적용

* refactor: 예상할 수 없는 예외 처리 및 어플리케이션 에러 로직 추가

* test: 예외시 500에러 400으로 변경

* fix: datasource url 수정

* fix: javax 의존 추가

* refactor: section의 순서 보장

* fix: sql문 Line -> line 변경

* feat: 배포 스크립트 생성

* refactor: line 삭제시 관련 구간도 모두 삭제되도록 수정

* refactor: sql문 외래키 표현 수정

* refactor: NotFoundException도 badRequest로 응답하도록 수정

* docs: 검증 요구사항 작성

* feat: 회원 기능 관련 검증 구현

* feat: 회원 기능 관련 검증 구현

* feat: 지하철 역 관련 기능 검증 구현

* feat: 지하철 노선 관리 기능 검증 구현

* feat: swagger 자동 문서화 기능 추가

* refactor: 패키지 내의 API만 문서로 보이도록 수정

* feat: 요금 조회 기능 구현

* feat: 노선 별 추가요금 구현

* feat: 연령별 요금 정책 추가

* test: 로그인시 나이에 맞는 적절한 요금 계산 TC 작성

* fix: rebase 했을 때, conflict 난 부분 수정

* refactor: AuthService 의 예외 처리 로직 리팩터링

* refactor: LineService 검증 로직 추가

* refactor: 불필요한 생성자 삭제

* refactor: 패키지명 변경

* refactor: StationService 검증 추가

* fix: 잘못된 존재 검증 수정

* feat: api 문서 추가

* fix: 문서가 생성되지 않는 버그 해결

* refactor: 사용자를 찾을 경우, Optional을 반환하도록 변경

* fix: typo 패스워트 -> 패스워드

* refactor: LoginMember를 넘기지 않고, age만 넘기도록 변경

* refactor: validation 에러메세지 가져오는 방식 변경

* refactor: logger 설정

* refactor: StringUtils 사용

* fix: bearer token 확인 로직 버그 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,40 +1,107 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.station.dao.StationDao;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.exception.DuplicateException;
+import wooteco.subway.exception.NotFoundException;
+import wooteco.subway.exception.SubwayException;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.infrastructure.dao.LineDao;
+import wooteco.subway.line.infrastructure.dao.SectionDao;
 import wooteco.subway.station.domain.Station;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.infrastructure.dao.StationDao;
+import wooteco.subway.station.ui.dto.LineResponse;
+import wooteco.subway.station.ui.dto.StationRequest;
+import wooteco.subway.station.ui.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationWithLinesResponse;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Map;
+import java.util.function.Function;
 
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toMap;
+
+@Transactional(readOnly = true)
 @Service
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
+    private final LineDao lineDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao, LineDao lineDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
+        this.lineDao = lineDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
+        validateThatStationExists(id);
         return stationDao.findById(id);
     }
 
-    public List<StationResponse> findAllStationResponses() {
-        List<Station> stations = stationDao.findAll();
-
-        return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+    private void validateThatStationExists(Long id) {
+        if (!stationDao.existsById(id)) {
+            throw new NotFoundException("존재하지 않는 역입니다.");
+        }
     }
 
+    public List<StationWithLinesResponse> findAllStationResponses() {
+        List<Station> allStations = stationDao.findAll();
+        List<Line> allLines = lineDao.findAll();
+
+        Map<Station, List<Line>> stationsWithLines = allStations.stream()
+            .collect(toMap(Function.identity(), station ->
+                allLines.stream()
+                    .filter(line -> line.getStations().contains(station))
+                    .collect(toList())
+            ));
+
+        return stationsWithLines.keySet().stream()
+            .map(station -> new StationWithLinesResponse(station.getId(), station.getName(),
+                convertLinesToLineResponses(stationsWithLines.get(station))))
+            .collect(toList());
+    }
+
+    private List<LineResponse> convertLinesToLineResponses(List<Line> lines) {
+        return lines.stream()
+            .map(line -> new LineResponse(line.getId(), line.getName(), line.getColor()))
+            .collect(toList());
+    }
+
+    @Transactional
     public void deleteStationById(Long id) {
+        validateThatStationExists(id);
+        validateDeletableStation(id);
         stationDao.deleteById(id);
     }
+
+    private void validateDeletableStation(Long id) {
+        if(sectionDao.existsByStationId(id)) {
+            throw new SubwayException("노선에 등록된 역은 삭제할 수 없습니다.");
+        }
+    }
+
+    @Transactional
+    public void updateStation(Long id, StationRequest stationRequest) {
+        validateStationRequest(id, stationRequest);
+        Station station = findStationById(id);
+        Station newStation = new Station(station.getId(), stationRequest.getName());
+
+        stationDao.updateStation(newStation);
+    }
+
+    private void validateStationRequest(Long id, StationRequest stationRequest) {
+        if (stationDao.existsByNameExceptId(id, stationRequest.getName())) {
+            throw new DuplicateException("이미 존재하는 역 이름입니다.");
+        }
+    }
+
 }
